Fix OptimizationConfig timeout_seconds type to support fractional seconds
Change timeout_seconds field type from int | None to float | None in OptimizationConfig TypedDict. This allows specifying sub-second timeouts (e.g., 0.3 seconds) for fast optimization runs in tests and time-sensitive applications. The None value continues to represent no timeout.

+++ libs/covenant_ml/src/covenant_ml/optimizer/types.py
@@ -252,7 +252,7 @@ class OptimizationConfig(TypedDict, total=True):
     """Configuration for an optimization run."""
 
     n_trials: int
-    timeout_seconds: int | None  # None = no timeout
+    timeout_seconds: float | None  # None = no timeout, float for fractional seconds
     n_startup_trials: int  # Random trials before TPE kicks in
     random_state: int
     direction: Literal["maximize", "minimize"]
